\name{pd.mirna.4.1}
\Rdversion{1.1}
\docType{package}
\alias{pd.mirna.4.1}
\title{
  Annotation package for pd.mirna.4.1.
}
\description{
  Annotation package for pd.mirna.4.1 built with pdInfoBuilder.
}
\details{
  This package is to be used in conjunction with the oligo package.
}
\keyword{package}
